requirements.txt
# No-Harm-VLLM Jury v3.0 Requirements

# Core dependencies (from v3.0)
requests>=2.31.0
PyYAML>=6.0

# vLLM inference (from v2.3)
numpy>=1.24.0
torch>=2.1.0
vllm>=0.6.0
transformers>=4.36.0
accelerate>=0.25.0

# HuggingFace (from v3.0)
huggingface_hub>=0.19.0
datasets>=2.14.0

# Visualization and reporting (from v2.3)
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scipy>=1.11.0

# Development (from v3.0)
pytest>=7.4.0
pytest-cov>=4.1.0
